Fix person_role_id and add obsserver to UserRating service

+++ src/main/java/org/solvd/recommendation/dao/IContentContributorDAO.java
@@ -6,5 +6,5 @@ import org.solvd.recommendation.util.CompositeKey3;
 import java.sql.SQLException;
 
 
-public interface IContentContributorDAO extends IDAO<ContentContributor, CompositeKey3<Long, Long, Integer>> {
+public interface IContentContributorDAO extends IDAO<ContentContributor, CompositeKey3<Long, Long, Long>> {
 }
